+++ src/bonuses/inspiration/inspiration-tenacious.mjs
@@ -1,0 +1,3 @@
+// https://www.d20pfsrd.com/classes/hybrid-classes/investigator/investigator-talents/paizo-investigator-talents/tenacious-inspiration-ex
+
+// When an investigator rolls his inspiration die, he can roll an additional inspiration die and take the higher result.