+++ src/bonuses/inspiration/inspiration-true.mjs
@@ -1,0 +1,5 @@
+// https://www.d20pfsrd.com/CLASSES/HYBRID-CLASSES/INVESTIGATOR/#TOC-True-Inspiration-Ex-:~:text=standard%20action.-,True%20Inspiration,-(Ex)
+
+// At 20th level, an investigator can use inspiration on all skill checks—even ones he isn’t trained in—and all ability checks without spending inspiration.
+//
+// In addition, whenever he expends inspiration on an ability check, attack roll, saving throw, or skill check, he adds 2d6 rather than 1d6 to the result. Some talents can affect this. If using the amazing inspiration investigator talent, he rolls 2d8 instead. If using this with empathy, tenacious inspiration, underworld inspiration, or a similar talent, he rolls two sets of inspiration dice and uses the higher of the two results.