+++ src/handlebars-handlers/enabled-label.mjs
@@ -1,5 +1,5 @@
 import { api } from '../util/api.mjs';
-import { localizeBonusLabel, localizeBonusTooltip } from '../util/localize.mjs';
+import { localize, localizeBonusLabel, localizeBonusTooltip } from '../util/localize.mjs';
 import { addNodeToRollBonus } from "./add-bonus-to-item-sheet.mjs";
 import { createTemplate, templates } from "./templates.mjs";
 
@@ -10,7 +10,9 @@ import { createTemplate, templates } from "./templates.mjs";
  * @param {string} args.key
  * @param {string} [args.label]
  * @param {HTMLElement} args.parent
+ * @param {string} [args.forLabel]
  * @param {string} [args.subLabel]
+ * @param {string} [args.text]
  * @param {string} [args.tooltip]
  * @param {object} options
  * @param {boolean} options.canEdit
@@ -19,24 +21,30 @@ import { createTemplate, templates } from "./templates.mjs";
 export function showEnabledLabel({
     item,
     key,
+    forLabel = '',
     label = '',
     journal,
     parent,
     subLabel = '',
+    text = '',
     tooltip = '',
 }, {
     canEdit,
     inputType,
 }) {
+    forLabel ||= key;
     label ||= localizeBonusLabel(key);
     tooltip ||= localizeBonusTooltip(key);
+    text ||= localize('PF1.Enabled');
     const div = createTemplate(
         templates.enabledLabel,
         {
+            for: forLabel,
             journal,
             label,
             parent,
             subLabel,
+            text,
             tooltip,
         },
     );