+++ src/util/get-skills.mjs
@@ -3,9 +3,12 @@ import { getSkillName } from './get-skill-name.mjs';
 
 /**
  * @param {Nullable<ActorPF>} actor
+ * @param {boolean} [isEditable] sanity for me to simplify places that need this code
  * @returns {Partial<Record<SkillId, string>>}
  */
-export const getSkills = (actor) => {
+export const getSkills = (actor, isEditable = true) => {
+    if (!isEditable) return {};
+
     /** @type {boolean} */
     const backgroundEnabled = !!game.settings.get('pf1', 'allowBackgroundSkills');